medicus-dictate/requirements.txt
faster-whisper>=1.0.0
sounddevice>=0.4.6
numpy>=1.24
pynput>=1.7.6
pyperclip>=1.8.2
pyautogui>=0.9.54
pystray>=0.19.5
Pillow>=10.0.0
tomli>=2.0.1 ; python_version < "3.11"
pyttsx3>=2.90

# Build-only (install when packaging a .exe):
# pyinstaller>=6.0
